+++ aiu/tags.py
@@ -6,3 +6,14 @@ TAG_GENRE = "genre"
 TAG_TITLE = "title"
 TAG_TRACK = "track"
 TAG_YEAR = "year"
+
+TAGS = frozenset([
+    TAG_ALBUM,
+    TAG_ALBUM_ARTIST,
+    TAG_ARTIST,
+    TAG_DURATION,
+    TAG_GENRE,
+    TAG_TITLE,
+    TAG_TRACK,
+    TAG_YEAR,
+])